feat: shared functions for some mutable element tests
