+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/impl/immutable/runImmutableEqualsTests.kt
@@ -1,12 +1,8 @@
 package xyz.lbres.kotlinutils.set.multiset.impl.immutable
 
-import xyz.lbres.kotlinutils.list.IntList
 import xyz.lbres.kotlinutils.set.multiset.* // ktlint-disable no-wildcard-imports no-unused-imports
 import xyz.lbres.kotlinutils.set.multiset.const.ConstMultiSetImpl
 import xyz.lbres.kotlinutils.set.multiset.const.ConstMutableMultiSet
-import xyz.lbres.kotlinutils.set.multiset.impl.MultiSetImpl
-import xyz.lbres.kotlinutils.set.multiset.testutils.runMultiSetMinusTests
-import xyz.lbres.kotlinutils.set.multiset.testutils.runMultiSetPlusTests
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 import kotlin.test.assertNotEquals
@@ -131,95 +127,4 @@ fun runImmutableEqualsTests() {
     mutableSet1.add(2)
     setSet2 = ConstMultiSetImpl(listOf(setOf(1, 3, 2)))
     assertEquals(setSet2, setSet1)
-}
-
-fun runImmutableMinusTests() {
-    runMultiSetMinusTests(
-        { MultiSetImpl(it) },
-        { MultiSetImpl(it) },
-        { MultiSetImpl(it) },
-        { MultiSetImpl(it) },
-        { MultiSetImpl(it) },
-        { ConstMutableMultiSet(it) }
-    )
-
-    // mutable elements
-    val mutableList1 = mutableListOf(1, 2, 3)
-    val mutableList2 = mutableListOf(0, 5, 7)
-    val listSet1: MultiSet<IntList> = multiSetOf(mutableList1, mutableList2)
-    val listSet2: MultiSet<List<Int>> = multiSetOf(listOf(1, 2, 3), listOf(0, 5, 7))
-
-    var expectedList: MultiSet<IntList> = multiSetOf()
-    assertEquals(expectedList, listSet1 - listSet2)
-    assertEquals(expectedList, listSet2 - listSet1)
-
-    mutableList2.add(3)
-    expectedList = multiSetOf(listOf(0, 5, 7, 3))
-    assertEquals(expectedList, listSet1 - listSet2)
-    expectedList = multiSetOf(listOf(0, 5, 7))
-    assertEquals(expectedList, listSet2 - listSet1)
-}
-
-fun runImmutablePlusTests() {
-    runMultiSetPlusTests(
-        { MultiSetImpl(it) },
-        { MultiSetImpl(it) },
-        { MultiSetImpl(it) },
-        { MultiSetImpl(it) },
-        { MultiSetImpl(it) },
-        { ConstMutableMultiSet(it) }
-    )
-
-    // mutable elements
-    val mutableList1 = mutableListOf(1, 2, 3)
-    val mutableList2 = mutableListOf(0, 5, 7)
-    val listSet1: MultiSet<IntList> = multiSetOf(mutableList1, mutableList2)
-    val listSet2: MultiSet<IntList> = multiSetOf(listOf(1), mutableList1)
-
-    var expectedList: MultiSet<IntList> = multiSetOf(listOf(1), listOf(1, 2, 3), listOf(1, 2, 3), listOf(0, 5, 7))
-    assertEquals(expectedList, listSet1 + listSet2)
-    assertEquals(expectedList, listSet2 + listSet1)
-
-    mutableList1.add(4)
-    expectedList = multiSetOf(listOf(1), listOf(1, 2, 3, 4), listOf(1, 2, 3, 4), listOf(0, 5, 7))
-    assertEquals(expectedList, listSet2 + listSet1)
-    assertEquals(expectedList, listSet1 + listSet2)
-}
-
-fun runImmutableIntersectTests() {
-    runMultiSetPlusTests(
-        { MultiSetImpl(it) },
-        { MultiSetImpl(it) },
-        { MultiSetImpl(it) },
-        { MultiSetImpl(it) },
-        { MultiSetImpl(it) },
-        { ConstMutableMultiSet(it) }
-    )
-
-    // mutable elements
-    val mutableList1 = mutableListOf(1, 2, 3)
-    val mutableList2 = mutableListOf(0, 5, 7)
-    val listSet1: MultiSet<IntList> = multiSetOf(mutableList1, mutableList2)
-    var listSet2: MultiSet<IntList> = multiSetOf(listOf(1), mutableList1)
-
-    var expectedList: MultiSet<IntList> = multiSetOf(listOf(1, 2, 3))
-    assertEquals(expectedList, listSet1 intersect listSet2)
-    assertEquals(expectedList, listSet2 intersect listSet1)
-
-    listSet2 = multiSetOf(listOf(1), listOf(1, 2, 3))
-    expectedList = multiSetOf(listOf(1, 2, 3))
-    assertEquals(expectedList, listSet1 intersect listSet2)
-    assertEquals(expectedList, listSet2 intersect listSet1)
-
-    mutableList2.clear()
-    mutableList2.add(1)
-    expectedList = multiSetOf(listOf(1), listOf(1, 2, 3))
-    assertEquals(expectedList, listSet1 intersect listSet2)
-    assertEquals(expectedList, listSet2 intersect listSet1)
-
-    mutableList1.add(4)
-    mutableList2.add(1)
-    expectedList = emptyMultiSet()
-    assertEquals(expectedList, listSet2 intersect listSet1)
-    assertEquals(expectedList, listSet1 intersect listSet2)
 }
